still need to fix bugs at d..

+++ src/ft_analyze_fmt.c
@@ -6,7 +6,7 @@
 /*   By: hokutosuzuki <hosuzuki@student.42toky      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/30 17:23:15 by hokutosuz         #+#    #+#             */
-/*   Updated: 2022/01/30 20:27:15 by hokutosuz        ###   ########.fr       */
+/*   Updated: 2022/01/31 14:34:57 by hokutosuz        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,13 +35,13 @@ void	ft_analyze_speci(const char *fmt, t_stock *lst,  size_t i)
 	else if (fmt[i] == 'p')
 		ft_print_address(lst, (size_t)va_arg(lst->args, void *));
 	else if (fmt[i] == 'd' || fmt[i] == 'i')
-		ft_print_decimal(lst, va_arg(lst->args, int));
+		ft_print_decimal(lst, va_arg(lst->args, long long));
 	else if (fmt[i] == 'u')
 		ft_print_unsigned(lst, va_arg(lst->args, unsigned int));
 	else if (fmt[i] == 'x')
-		ft_print_hex(lst, va_arg(lst->args, unsigned int));
+		ft_print_hex(lst, va_arg(lst->args, int));
 	else if (fmt[i] == 'X')
-		ft_print_hex_cap(lst, va_arg(lst->args, unsigned int));
+		ft_print_hex_cap(lst, va_arg(lst->args, int));
 	else if (fmt[i] == '%')
 	{
 		lst->total_len++;
